requirements.txt
# Universal Markdown Audio Transcription System Requirements

# Core transcription engine
openai-whisper>=20231117

# Configuration support
PyYAML>=6.0

# Audio/video processing dependencies (automatically installed with whisper)
torch>=2.0.0
torchaudio>=2.0.0
ffmpeg-python>=0.2.0

# Optional: For advanced audio processing
# librosa>=0.10.0
# soundfile>=0.12.0